finish service for daily checkup port and models
